Lost the thread of microchanges entirely today. Here is a second attempt at cleaning up the demo
